Mutual Information score

The function to compute collocation association measure with Mutual Information.

collex_MI(df)

Arguments

df
The output of assoc_prepare.

Value

A tibble consisting of the collocates (column w), co-occurrence frequencies with the node (column a), the expected co-occurrence frequencies with the node (column a_exp), the direction of the association (e.g., attraction or repulsion) (column assoc), the Mutual Information score (column MI), and two uni-directional association measures of Delta P.

Examples

out <- colloc_leipzig(leipzig_corpus_list = demo_corpus_leipzig,
    pattern = "ke", # it is a preposition meaning 'to(wards)'
    window = "r",
    span = 2L,
    save_interim = FALSE)

#> Detecting a 'named list' input!

#> You chose NOT to SAVE INTERIM RESULTS, which will be stored as a list in console!

#> 1. Tokenising the "ind_mixed_2012_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_mixed_2012_1M.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind_news_2008_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_news_2008_300K.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind_news_2009_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_news_2009_300K.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind_news_2010_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_news_2010_300K.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind_news_2011_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_news_2011_300K.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind_news_2012_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_news_2012_300K.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind_newscrawl_2011_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_newscrawl_2011_1M.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind_newscrawl_2012_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_newscrawl_2012_1M.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind_newscrawl_2015_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_newscrawl_2015_300K.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind_newscrawl_2016_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_newscrawl_2016_1M.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind_web_2011_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_web_2011_300K.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind_web_2012_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_web_2012_1M.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind_wikipedia_2016_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind_wikipedia_2016_1M.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind-id_web_2013_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind-id_web_2013_1M.

#> 2.1 Gathering the collocates for 'ke' ...

#> 1. Tokenising the "ind-id_web_2015_3M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'ke' in ind-id_web_2015_3M.

#> 2.1 Gathering the collocates for 'ke' ...

#> 3. Storing all of the outputs...

#> 
#> DONE!

assoc_tb <- assoc_prepare(colloc_out = out,
     stopword_list = collogetr::stopwords[collogetr::stopwords != "ke"])

#> Your colloc_leipzig output is stored as list!

#> You chose to combine the collocational and frequency list data from ALL CORPORA!

#> Tallying frequency list of all words in ALL CORPORA!

#> You chose to remove stopwords!

collex_MI(assoc_tb)

#> # A tibble: 301 x 8
#> # Groups:   node, w [301]
#>    w        node      a a_exp assoc        MI dP_collex_cue_cxn dP_cxn_cue_coll…
#>    <chr>    <chr> <int> <dbl> <chr>     <dbl>             <dbl>            <dbl>
#>  1 rumah    ke       10 0.578 attracti…  4.11           0.0361            0.104 
#>  2 luar     ke        6 0.273 attracti…  4.46           0.0219            0.133 
#>  3 arah     ke        5 0.127 attracti…  5.30           0.0186            0.244 
#>  4 berbagai ke        5 0.33  attracti…  3.92           0.0179            0.0899
#>  5 kata     ke        5 1.33  attracti…  1.91           0.0141            0.0177
#>  6 negara   ke        5 0.647 attracti…  2.95           0.0167            0.0428
#>  7 daerah   ke        4 0.489 attracti…  3.03           0.0134            0.0457
#>  8 tempat   ke        4 0.317 attracti…  3.66           0.0141            0.0737
#>  9 belakang ke        3 0.076 attracti…  5.30           0.0112            0.244 
#> 10 dua      ke        3 0.609 attracti…  2.30           0.00915           0.0250
#> # … with 291 more rows