Sentence match retriever

The function extract full sentence-match for a (set of significant) collocate(s) for a given nodeword.

colloc_sentmatch(
  collout,
  colloc = NULL,
  wspan = NULL,
  nodeword = NULL,
  sampled = NULL
)

Arguments

collout
List output of colloc_leipzig.

colloc
Character vector of the collocate(s) whose sentence match(es) to be retrieved.

wspan
Character vector of the window span in which the collocates occur. Default to NULL, which will retrieve the collocate's occurrence in all span.

nodeword
Character vector specifying one of the nodewords if search parameter in colloc_leipzig includes more than one nodeword.

sampled
Integer vector indicating the number of random sample of the sentence match to be retrieve. Default to NULL, which will retrieve all sentence-matches.

Value

Character vector of sentence-match(es).

Details

See also

Examples

collout <- colloc_leipzig(leipzig_corpus_list = demo_corpus_leipzig,
                     pattern = "mengatakan",
                     window = "r",
                     span = 3,
                     save_interim = FALSE)

#> Detecting a 'named list' input!

#> You chose NOT to SAVE INTERIM RESULTS, which will be stored as a list in console!

#> 1. Tokenising the "ind_mixed_2012_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_mixed_2012_1M.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind_news_2008_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_news_2008_300K.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind_news_2009_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_news_2009_300K.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind_news_2010_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_news_2010_300K.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind_news_2011_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_news_2011_300K.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind_news_2012_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_news_2012_300K.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind_newscrawl_2011_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_newscrawl_2011_1M.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind_newscrawl_2012_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_newscrawl_2012_1M.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind_newscrawl_2015_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_newscrawl_2015_300K.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind_newscrawl_2016_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_newscrawl_2016_1M.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind_web_2011_300K" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_web_2011_300K.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind_web_2012_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_web_2012_1M.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind_wikipedia_2016_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind_wikipedia_2016_1M.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind-id_web_2013_1M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind-id_web_2013_1M.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 1. Tokenising the "ind-id_web_2015_3M" corpus. This process may take a while!

#>     1.1 Removing one-character tokens...

#>     1.2 Lowercasing the tokenised corpus...

#>     At least a match is detected for 'mengatakan' in ind-id_web_2015_3M.

#> 2.1 Gathering the collocates for 'mengatakan' ...

#> 3. Storing all of the outputs...

#> 
#> DONE!

colloc_sentmatch(collout,
                 colloc = "bahwa",
                 sampled = 10)

#> Warning: Length of logical index must be 1 or 451, not 0

#> Returning all matches!
#> Length of matches (0) is lower than the number of the queried sample (10).

#> 

#> character(0)

# This will produce message indicating that
# the queried sample number is higher than
# the sentence match for "akan"
colloc_sentmatch(collout,
                 colloc = "akan",
                 sampled = 10)

#> Warning: Length of logical index must be 1 or 451, not 0

#> Returning all matches!
#> Length of matches (0) is lower than the number of the queried sample (10).

#> 

#> character(0)