Compute Log10 Odds Ratio

internal function called by collex_logOR to compute Log10 Odds Ratio (Gries 2014).

logOR_compute(a, b, c, d)

Arguments

a
co-occurrence frequency between a given collocate and the nodeword.

b
co-occurrence frequency between the collocate and other words.

c
co-occurrence frequency between the nodeword and the other words.

d
co-occurrence frequency between all other words.

Value

A double-numeric value of Log10 Odds Ratio score

References

Rajeg, Gede Primahadi Wijaya & I Made Rajeg. 2019. Analisis Koleksem Khas dan potensinya untuk kajian kemiripan makna konstruksional dalam Bahasa Indonesia. In I Nengah Sudipa (ed.), ETIKA BAHASA Buku persembahan menapaki usia pensiun: I Ketut Tika, vol. 1, 65–83. Denpasar, Bali, Indonesia: Swasta Nulus. https://doi.org10.26180/5bf4e49ea1582. https://osf.io/preprints/inarxiv/uwzts/ (30 January, 2019).

Gries, S. T. (2014). Coll.analysis 3.5. A script for R to compute perform collostructional analyses. http://www.linguistics.ucsb.edu/faculty/stgries/teaching/groningen/index.html.

Examples

# Data from Table 2 in Rajeg and Rajeg (2019:71)
a <- 9
b <- 1
c <- 408
d <- 305

logOR_compute(a, b, c, d)

#> [1] 0.6754515

# the above data produces positive log OR
# Observed freq. for cell "a" is larger than the Expected freq.

# Data from Table 3 in Rajeg and Rajeg (2019:74)
a <- 1
b <- 6
c <- 416
d <- 300

logOR_compute(a, b, c, d)

#> [1] -0.7785926

# the above data produces negative log OR
# Expected freq. for cell "a" is larger than the Observed freq.